Expanded debug mode in Makefile
This adds switches for *debug mode*, for more strict testing of the code and for enabling debugging.

In particular, `-Wall` has been added. Changes have been made to the code so that all warnings are handled.

Further:

- Added target `run_all` to the Makefile. This is intended to be an internal target, for running all the examples.
  The goal is to check if there are any errors during runtime when running any example.
- Target `help` is now at the top of the makefile, so that it is guaranteed to run when just `make` is invoked.
  In addition, it is now one of the first things you see when you open the makefile.
- In the makefile, there is now a distinction between `TOOLS` and `EXAMPLES`.

+++ Lib/VideoCore/Mailbox.cpp
@@ -93,7 +93,7 @@ static int mbox_property(int file_desc, void *buf)
 
 #ifdef DEBUG
    unsigned *p = (unsigned*) buf; int i; unsigned size = *(unsigned *)buf;
-    for (i=0; i<size/4; i++)
+   for (i=0; i< (int) size/4; i++)
        printf("%04x: 0x%08x\n", i * (unsigned) sizeof(*p), p[i]);
 #endif  // DEBUG
